==> 8_Manuscript.Rmd <==
---
title: Bumble bee probability of occurrence responds to interactions between local and landscape land use, climatic niche properties and climate change
author: "Tim Newbold^1,\\*^, Jeremy Kerr^2^, Peter Soroye^2,&dagger;^, Jessica J. Williams^1,&Dagger;^"
output:
  bookdown::word_document2:
    toc: false
    fig_caption: TRUE
    number_sections: no
always_allow_html: true
csl: ecology-letters.csl
bibliography: bumblebees_lu_cc_refs.bib
---

^1^Centre for Biodiversity and Environment Research, Department of Genetics, Evolution and Environment, University College London, London, UK

^2^Department of Biology, University of Ottawa, Ottawa, Ontario, Canada

^&dagger;^Current address: Wildlife Conservation Society Canada, Toronto, Ontario, Canada

^&Dagger;^Current address: Department of Life Sciences, Silwood Park Campus, Imperial College London, London, UK

^\*^Correspondence: Tim Newbold, Centre for Biodiversity & Environment Research, Department of Genetics, Evolution & Environment, University College London, Gower Street, London WC1E 6BT, United Kingdom; Email: [t.newbold@ucl.ac.uk](mailto:t.newbold@ucl.ac.uk); Tel: +44(0)2031087569

**Short title:** Land-use-climate interactions and bumble bees

**Keywords:** Bumble bees, Climate change, Fertilizer, Land use, Natural habitat, Probability of occurrence, Pesticide, Pressure interactions, Thermal niche

**Type of article:** Letter

**Authorship statement:** TN, JK, PS and JJW designed the analyses. TN performed all main analyses. TN and PS provided the methods for quantifying species' thermal niche position. TN wrote the first draft of the manuscript. All authors contributed to editing the manuscript. TN and JK secured the funding to support the presented research.

**Number of words in abstract:** 150

**Number of words in main text (excluding abstract, acknowledgments, references, and table and figure legends):** 4,999

**Number of references:** 69

**Number of figures:** 3

**Number of tables:** 0

<!-- SETUP: Global options -->
```{r setup,echo=FALSE,results=FALSE}
knitr::opts_chunk$set
options(scipen=20)
```

\newpage

# Abstract

Insect biodiversity is changing rapidly, driven by a suite of pressures, notably land use, land-use intensification, and increasingly climate change. We lack large-scale evidence on how land use and climate change interact to drive insect biodiversity changes. We assess bumble bee responses to interactive effects of land use and climate pressures across North America and Europe. Probability of occurrence increases in landscapes with a higher proportion of natural habitat and a shorter history of human disturbance. Responses to climate warming relative to historical conditions are weakly negative in natural habitats but positive in human land uses, while human land use reduces probability of occurrence most in the centre of species’ temperature niches. We estimate that the combined pressures have reduced bumble bee probability of occurrence by 45% across sampled natural habitats, and 56% across human land uses, highlighting the pervasive influence that human pressures have had on biodiversity across habitats.

\newpage

# Introduction

Insect communities are changing rapidly, although reported biodiversity trends differ in direction and magnitude. Several studies have detected steep declines in the abundance, species richness and distributions of terrestrial insects [@Kerr2015; @Lister2018; @vanKlink2020; @Janousek2023], while other studies report little overall change in insect biodiversity on average [@Outhwaite2020; @Macgregor2019].

Human land use and land-use intensification are key drivers of insect biodiversity changes. For bees, conversion of natural habitats to agriculture and other human uses often has negative effects [@DePalma2017; @Evans2018; @Samuelson2018; @Janousek2023], although responses vary strongly among species [@Prestele2021; @Cariveau2013] and regions [@DePalma2016]. Within agricultural areas, intensification of farming practices (e.g. lowering of crop diversity, or removal of flower-rich field margins) is associated with further reductions, on average [@Hemberger2021; @Kennedy2013]. One facet of agricultural intensification that is particularly consequential for bees is the application of chemical pesticides, with toxicity to bees often increasing over time [@Douglas2020], leading to reductions in bee distributions, population persistence, and reproductive performance [@Whitehorn2012; @Woodcock2016; @Siviter2021; @Stuligross2021; @Nicholson2024]. 

The biodiversity impacts of habitat modifications can operate over wider landscapes and longer timescales than captured by localized biodiversity samples. Natural habitat availability in the landscape is important for maintaining bee diversity, including in agricultural habitats [@Evans2018; @Mola2021; @Rutschmann2022; @Proesmans2019; @GutierrezChacon2018]. However, responses often vary among species and types of natural habitat (e.g., forest versus grassland) [@Kammerer2021; @Goulson2010; @GutierrezChacon2018], and are sometimes absent or negative, for example when farmed areas provide abundant floral resources [@Christman2022; @Mola2021]. Farming practices may impact biodiversity within adjacent natural habitats, for example, through drift of pesticides and other agricultural chemicals out of the farmlands to which they are applied [@Krupke2012]. Over time, agricultural intensity often increases, with loss of remnant natural habitats [@Tscharntke2005], effects which are associated with reductions in bumble bee occurrence [@Hemberger2021].

Climate change exerts a strong influence on bees and other insects, and its effects are growing rapidly. Already, bumble bee distributions have been contracting at southern range margins [@Kerr2015] and moving to higher elevations [@Pyke2016], while the Community Temperature Index of communities (the average temperature affiliation of species in a community) has been increasing [@Fourcade2019; @Hemberger2024], consistent with expectations under climate warming. Indeed, analyses of temporal changes in bee diversity have revealed an important role for warming temperatures in observed declines [@Kammerer2021; @Soroye2020; @Janousek2023]. Future projections suggest that climate change impacts on bumble bees are likely to increase [@Prestele2021; @Marshall2018].

Recent evidence points toward interactive effects of land use, land-use intensification and climate change on biodiversity, but fewer studies have focused on insects than on the better-studied vertebrates (e.g., [@Hamblin2017; @Oliver2017; @Fourcade2019; @Waldock2020; @Kammerer2021; @Outhwaite2022; @Janousek2023]). Such interactive effects are driven by two key mechanisms: habitat disturbance impeding species' range shifts in response to climate change [@Oliver2014]; and local climatic changes associated with conversion of natural habitats [@Oliver2014; @Williams2020]. Insect species tolerant of warmer and drier conditions are, on average, more likely to occur within agricultural and urban land uses than other species [@Oliver2017; @Hamblin2017; @Waldock2020], probably resulting from altered local climatic conditions in these areas. Studies on vertebrates have revealed that populations living close to the upper realized thermal niche limit of species respond more negatively to human land uses than other populations [@Williams2021].

We present a continental-scale study, encompassing Europe and North America, considering the interacting effects of climate change, land use and land-use intensification on the average occurrence probability of bumble bees. A number of studies have investigated large-scale (nation or continent-wide) effects of climate and land-use variables on bees. Notably, [@Fourcade2019] assessed change in bumble bee species richness in Norway as a result of climate and land-cover changes, [@Kammerer2021] tested the effect on bee species richness of climate, land-cover and insecticide toxic load in three US states, [@Janousek2023] showed climate change, land cover and pesticide application impacts on the western bumble bee across the US, and [@Hemberger2024] documented restructuring of bumble bee assemblages in the US favouring species tolerant of higher temperatures. Here, we make three further developments over previous work. First, we use a recent characterization of the realized thermal niche of bumble bees [@Soroye2020] to test the effect of climate change dependent upon the position of a population within species' realized thermal niche, allowing us to go beyond recent work assessing general interactive effects of land use and climate change across all insects [@Outhwaite2022]. Second, we use estimates of land-use history to test whether bumble bee occurrence depends on the time since landscapes were converted to human use, in conjunction with present land uses, habitat extent, and pesticide toxicity. We expect that bumble bee occurrence probability will be lower in agricultural areas, particularly where land-use intensity is higher, where landscapes have been dominated by human activities for longer, and where less natural habitat remains within the landscape. We further predict that occurrence probability will decline most strongly amid agricultural land uses where bumble bee populations are near their upper thermal limits, and especially where temperature has recently increased. Third, we use the models to infer how bumble bee occurrence probability within sampled habitats (both natural and human-modified habitats) has been shaped by the pressures currently acting on them.

# Material and Methods

## Data on Responses to Land Use

<!-- SUMMARY STATISTICS: data summaries -->
```{r summarize_data,echo=FALSE,warning=FALSE,results=FALSE}
# Set input directories
dataDir <- "2_PrepareDiversityData/"
modelsDir <- "3_RunModels/"

# Load the compiled species-level data
diversity <- readRDS(paste0(dataDir,"diversity_data.rds"))

allData <- readRDS(paste0(modelsDir,"ModelData.Rds"))

sites <- readRDS(paste0(modelsDir,"AnalysisSites.rds"))

n.studies <- format(length(unique(allData$SS)),big.mark=",",scientific=FALSE)
n.records <- format(nrow(allData),big.mark=",",scientific=FALSE)
n.species <- format(length(unique(allData$Best_guess_binomial)),big.mark=",",
                    scientific=FALSE)
n.countries <- format(length(unique(allData$Country)),big.mark=",",
                      scientific=FALSE)

countries <- unique(allData$Country)

# Put countries in more logical order
countries2 <- na.omit(countries[c(5,1,6,4,10,7,11,13,8,12,9,2,3)])

# Create text string
countries2 <- paste(countries2,collapse=", ")

# Replace last comma with 'and'
countries2 <- sub(",([^,]*)$", " and\\1", countries2)

n.sites <- format(nrow(sites),big.mark=",",scientific=FALSE)

# Get sample years
diversity$SampleYearMid <- as.integer(
  sapply(X = paste(diversity$Sample_midpoint),
         FUN = function(x) strsplit(x,'-')[[1]][1]))

first.year <- min(diversity$SampleYearMid)
last.year <- max(diversity$SampleYearMid)
```

Data describing bumble bee occurrence (recorded presences and absences) across land-use types were derived from the PREDICTS database, which collates biodiversity records from individual studies that conducted snapshot spatial samples of biodiversity along land-use or land-use-intensity gradients [@Hudson2017]. Studies included in the database had to meet four criteria: 1) sampling methods were published; 2) biodiversity was sampled at more than one location; 3) sampled locations spanned a gradient of land-use type and/or intensity; and 4) the sampling protocol was the same across sampled sites. Any records without geographical coordinates were omitted from our analysis.

The final dataset was derived from `r n.studies` original studies (listed in the Supplementary Material), and consisted of `r n.records` records for `r n.species` bumble bee species, sampled at `r n.sites` locations in `r n.countries` countries in North America, and Europe (Figure S1; Table S1). The original samples collated in the PREDICTS database were recorded in the field between `r first.year` and `r last.year`. Sampled species were those assessed as being of lower risk of extinction (at least for European species; Figure S2), which are expected to respond less to environmental changes, thus likely rendering inferred effects of climate change and land use conservative.

## Explanatory Variables

We considered local land use by dividing sampled sites into natural or human-modified habitats. In the PREDICTS database, land-use type is classified, based on the habitat descriptions provided by the original authors, into six broad categories: primary vegetation (natural habitat with no record of past destruction), secondary vegetation (habitat recovering to its natural state after past destruction by human actions or extreme natural events), plantation forests (areas used to cultivate woody crops), cropland (areas used to cultivate herbaceous crops, including fodder for livestock), pasture (areas used regularly or permanently to graze livestock), and urban (areas of human settlement, buildings, or managed for recreation). We treated primary and secondary vegetation as natural habitats, and all other land-use types as human-modified. Although bumble bees may respond more strongly to landscape habitat availability than to local land-use conditions [@Janousek2023], the latter can be important [@Evans2018], and are considered here also because they are fundamental to the sampling structure of the PREDICTS database [@Hudson2014].

We investigated how surrounding habitat conditions, cumulative pesticide toxicity, and species' thermal niche properties shape bumble bee occurrence in both natural and human-modified habitats. Information on habitat condition in the landscapes surrounding sampled sites comprised estimates of natural habitat availability, and the length of time the habitat has been substantially modified by humans. We obtained estimates of the percentage of natural habitat in the 2 &times; 2-km grid cell within which biodiversity was sampled (we tested the robustness of our results to using 1-km and 5-km grid cells; Supplementary Material, Figures S7 & S8). Original mapped estimates of different land-use types were from a down-scaled land-use projection for 2005 at 30-arc-second spatial resolution [@Hoskins2016]. These land-use maps estimate the fraction of each grid cell covered by five out of the six land-use types recognised in the PREDICTS database (not plantation forests), and were evaluated against the land-use classifications in the PREDICTS database [@Hoskins2016]. We summed the primary and secondary vegetation maps to obtain estimates of natural habitat. We projected and resampled this map of fractional cover of natural habitats to an equal-area grid (Behrmann projection) at 1-km resolution, using bilinear interpolation (using the terra R package Version 1.7-71; [@Hijmans2024]). All data-processing code was implemented in R Version 4.3.1 [@R431]. We then aggregated the resulting 1-km map by factors of two or five, to obtain the maps at 2-km and 5-km grain, calculating the mean fractional natural-habitat cover. Although there is a tendency for the percentage of surrounding natural habitat to be lower for human-modified land uses compared to natural habitats, this association is weak (ANOVA R^2^ = 0.06 at 2-km grain). We derived estimates of the duration of substantial human habitat modification from the land-use harmonisation project land-use reconstruction at 0.5° spatial resolution [@Hurtt2011], which estimates the fractional cover of primary vegetation, secondary vegetation, croplands, pastures and urban areas from 1500 to 2005. The HYDE historical land-use model, from which the reconstructed land-use estimates are interpolated has an interval of one century until 1700, and decadal thereafter [@KleinGoldewijk2011]. Thus, longer estimates of the duration of landscape modification are subject to a higher degree of uncertainty. We measured the duration of substantial human habitat modification as the number of years since a 0.5° grid cell was first 30% converted to human-modified land-use types (croplands, pastures or urban areas). Above 30% loss of natural habitats from landscapes, many of the most sensitive bird and mammal species are substantially negatively impacted [@Andren1994]. The appropriate threshold for insect species (including bumble bees) is unclear. Therefore, we tested the sensitivity of using conversion thresholds of 10% or 50% (Figures S9 & S10). Grid cells that did not reach a given conversion threshold were considered to have a substantial modification duration of 0 years.

To calculate pesticide toxicity, we used modelled estimates of application density (kg/ha) of the 20 dominant pesticide active ingredients for each of 6 individual crops (corn, soyabean, wheat, cotton, rice, and alfalfa) or 4 crop groups (vegetables and fruit, orchards and grapes, pasture and hay, and other crops) &ndash; totalling 95 different active ingredients across all crops [@Maggi2019]. These data are presented at 5-arc-minutes spatial resolution. We used the low rather than high modelled estimates in our main models, but tested the robustness of our models to using the high estimates (Figure S11). Estimates of application density for each pesticide active ingredient were divided by estimates of LD50 for that chemical, derived from the Pesticide Properties Database [@Lewis2016], using estimates for honey bees, since records of toxicity for bumble bees are incomplete. We used estimates of contact LD50s only (available for 87/95 active ingredients). We summed the toxicity estimates across the 20 active ingredients for each of the 10 crops/crop groups (200 toxicity maps in total).

For species' thermal niche properties, we used previously published methods and associated records of bumble bee occurrence [@Soroye2020]. Specifically, we estimated: 1) the position of a population within the species' realized thermal niche, by estimating how the temperatures experienced by the population relate to the realized thermal niche of the species in a baseline period before the onset of rapid climate changes (1901-1975); and 2) how changes in temperature between the baseline period and a recent period (2000-2014) have shifted the position of the population within this realized thermal niche. Realized thermal niche limits were estimated for each species as the mean of the five lowest minimum monthly temperatures and the mean of the five highest maximum monthly temperatures across all bumble bee spatial records in the baseline period (from an extensive database of species occurrence records across Europe and North America; see [@Soroye2020] for details). We then derived the maximum monthly temperatures for all twelve months in each baseline or recent year, and rescaled them such that a value of 0 equates to the minimum realized thermal niche limit of the species, and a value of 1 to the maximum niche limit. We then averaged the 12 (rescaled) monthly values to derive the final mean niche position for each year. Finally, we averaged estimates across all years in either the baseline or recent period. A thermal niche position value of 0 represents a population at the lower limit of the species’ realized thermal niche, while a value of 1 represents a population at the upper limit of the species’ realized thermal niche. Monthly minimum and maximum temperature estimates for the period 1901 to 2015 were obtained from the CRU global gridded climate reconstruction at 0.5° spatial resolution, Version 3.24.01 [@CRU32401], which was the most recent version of this dataset available at the time of calculation [@Soroye2020]. Because our study area encompasses regions well covered by weather stations, the CRU estimates of temperature for the sites in our analysis were all derived from data from a minimum of 22 weather stations at the beginning of the baseline period (January 1901), and 31 at the end of the recent period (December 2014). 

To control for potential confounding effects of elevation, we included in the model estimates of elevation at 30-arc-second spatial resolution, from WorldClim Version 2.1 [@Fick2017]. All spatial data manipulation was carried out using the terra R package Version 1.7-71 [@Hijmans2024].

## Statistical Analysis

All analysis code was implemented in R Version 4.3.1 [@R431]. We used a binomial Bayesian hierarchical model to fit species' presence or absence. Species were considered absent if they were not recorded but had been targeted by a particular study (i.e., they were recorded in at least one other sampled site within a study). As random effects, we fitted: 1) Study identity, to control for differences in sampling protocols and broad geographic differences; 2) Site identity nested within study, to control for site-specific factors; and 3) Species identity, to account for variation in responses among species. We were unable to account for variations in sampling effort beyond incorporating the random effect of study identity, because the bumble bee data in the PREDICTS database were collected using many different sampling methods, with non-comparable measures of sampling effort. Although there may be important differences in responses between the two continents, the dataset was too small to include such variation in the models. As a check of the robustness of our results, we also fit a zero-inflated negative binomial model of relative abundance (Figure S12). Similarly, while there may be interspecific variation in responses, there was not sufficient sampling of each species to allow fitting of random slopes in the models. As fixed effects, we considered local land-use type, surrounding natural habitat availability, the duration of substantial human modification of landscapes, cumulative pesticide toxicity, baseline position within species' thermal niche, change in thermal niche position between the baseline and recent periods, and elevation. To avoid model over-fitting, given the small size of the dataset, we fitted linear effects for all continuous variables, except for baseline thermal niche position, for which we fit a quadratic polynomial. Physiological experiments have long established that, for any given species, performance tends to be maximal at intermediate temperatures, and decline at colder and hotter temperatures [@Pawar2024]. Cumulative pesticide toxicity and elevation were log~e~-transformed to deal with the right-skewed distribution of values, first adding a value of 1 to pesticide toxicity and 2 to elevation to deal with zero and (for elevation) negative values. We considered interactions between local land-use type and the other continuous variables. We also explored interactions with surrounding natural habitat availability, but estimated coefficients for these interactions were subject to a high degree of uncertainty, probably owing to slight correlations among the different landscape-level variables. The structure of the final model was:

$$Occurrence \sim LandUse + NaturalHabitat + PesticideToxicity + LandscapeModificationDuration + poly(BaselineThermalNichePosition,2) + DeltaThermalNichePosition + LandUse:NaturalHabitat + LandUse:PesticideToxicity + LandUse:LandscapeModificationDuration + LandUse:poly(BaselineThermalNichePosition,2) + LandUse:DeltaThermalNichePosition + Elevation + (1|SS) + (1|SSBS) + (1|TaxonName)$$

Where 'poly(BaselineThermalPosition,2)' denotes the fitting of a quadratic polynomial term, 'SS' denotes study identity, and 'SSBS' denotes site identity (nested within study). All models were implemented in the brms R package Version 2.21.0 [@Burkner2021]. All parameters took the default, weakly informative priors. We ran 4 chains, and assessed convergence by inspection of Rhat values. Chains were run with a burn-in of 1,000 iterations and sampling period of 1,000 iterations. For comparison, we also built a model containing just land use as a fixed effect. Model explanatory power was assessed using a version of R^2^ values designed for Bayesian models [@Gelman2019], calculated using the bayes_R2 function (brms R package). Collinearity among the explanatory variables was sufficiently low to prevent biases (Pearson's correlation coefficient, r < 0.55). We tested for spatial autocorrelation across all of the residuals of the final model, and in the residuals associated with each underlying study using Moran’s tests, implemented in the spdep R package Version 1.2-8.

To reveal the impacts that cumulative pressures have had on bumble bee occurrence probability across natural and human-modified habitats, we used our model to infer biodiversity across the sampled sites, based on the values of the explanatory variables at the time of bumble bee sampling, compared to model predictions under a reference condition in which we assumed that landscapes were entirely composed of natural habitat, with no history of human modification, subject to no application of pesticides, and with no change in climate. Although this reference is entirely hypothetical, not reflecting actual conditions at any time in recent history, it nevertheless serves to illustrate the likely magnitude of human activities on bumble bee assemblages across natural and human-modified habitats. To derive model-estimated differences between reference and actual conditions, we drew 1,000 sets of coefficient estimates from the posterior samples of the model. We then calculated estimates of the percentage difference in average occurrence probability between reference and actual environmental conditions for each of these sets of coefficient estimates.

# Results

<!-- PLOTTING: Setup for model results plots -->
```{r results_setup, echo=FALSE,warning=FALSE}
#### Model results plot - Setup ####

# Set input directory
inDir <- "3_RunModels/"

# Load required packages
suppressMessages(suppressWarnings(library(terra))) # Version 1.7-71
suppressMessages(suppressWarnings(library(brms))) # Version 2.21.0
suppressMessages(suppressWarnings(library(Hmisc))) # Version 5.1-1
suppressMessages(suppressWarnings(library(ggplot2))) # Version 3.5.1
suppressMessages(suppressWarnings(library(ggstance))) # Version 0.3.7
suppressMessages(suppressWarnings(library(cowplot))) # Version 1.1.3
suppressMessages(suppressWarnings(library(sf))) # Version 1.0-19
suppressMessages(suppressWarnings(library(snow))) # Version 0.4-4
suppressMessages(suppressWarnings(library(parallel))) # Version 4.3.1
suppressMessages(suppressWarnings(library(spdep))) # Version 1.2-8

# Log-transform elevation, which is very skewed
diversity$LogElevation <- log(diversity$Elevation+2)

# Log-transform pesticide application, which is also very skewed
diversity$LogPestToxLow <- log(diversity$PestToxLow+1)

diversity$UI <- paste(diversity$LandUse,diversity$Use_intensity,sep="_")
diversity$UI[grepl("Natural",diversity$UI)] <- "Natural"
diversity$UI[grepl("NA",diversity$UI)] <- NA
diversity$UI[grepl("Cannot decide",diversity$UI)] <- NA
diversity$UI <- factor(diversity$UI)
diversity$UI <- relevel(diversity$UI,ref="Natural")

# Select relevant columns
modelData <- diversity[,c('occur','LandUse','TEI_BL',
                          'TEI_delta','LogElevation','SS','SSBS',
                          'Taxon_name_entered','Longitude','Latitude',
                          'Best_guess_binomial','Country',
                          'LogPestToxLow',
                          'NaturalHabitat2k','AgeConv')]

# Remove rows with any NA values
modelData <- na.omit(modelData)

# Load the final model
finalModel <- readRDS(paste0(inDir,"FinalModelBRMS.rds"))
```

<!-- STATISTICS: Model statistics -->
```{r calc_stats,echo=FALSE,warning=FALSE}

# Load the final model with just land use among the main
# explanatory variables
finalModelLU <- readRDS(paste0(inDir,"FinalModelLandUseOnly.rds"))

# Get Bayesian pseudo-R2 values for the final model
r2.marginal <- round(bayes_R2(finalModel,re.form=NA)[,'Estimate'],2)
r2.conditional <- round(bayes_R2(finalModel)[,'Estimate'],2)

# Get the pseudo-r-squared values of the land-use-only model
r2.lu.mod.cond <- round(bayes_R2(finalModelLU)[,'Estimate'],2)
r2.lu.mod.marg <- round(bayes_R2(finalModelLU,re.form=NA)
                        [,'Estimate'],3)

# Get the maximum Rhat value for the estimated coefficients
max.rhat <- round(max(summary(finalModel)$fixed$Rhat),2)
```

Landscape availability of natural habitat, duration of substantial human habitat modification, cumulative pesticide toxicity, thermal niche properties and climate change all contributed to explaining bumble bee occurrence probability, with effects strongly dependent on local land-use type (Figure \@ref(fig:figforest)). A model containing just land-use type explained little variation in bumble bee occurrence (R^2^~marginal~ = `r r2.lu.mod.marg`; R^2^~conditional~ = `r r2.lu.mod.cond`). The full model was substantially better (R^2^~marginal~ = `r r2.marginal`; R^2^~conditional~ = `r r2.conditional`). In the full model, land-use type had a negligible effect on bumble bee probability of occurrence, but played an important role in interaction with the other variables.

<!-- PLOTTING: Forest plot -->
```{r figforest, echo=FALSE, warning=FALSE, fig.width=17.5/2.54, fig.height=12/2.54,out.width="100%", dpi=1200, fig.cap="**Forest plot of coefficient estimates from the full binomial Bayesian hierarchical model relating bumble bee occurrence to land use, landscape habitat, pesticide toxicity and thermal niche properties**. Points represent median coefficient estimates, thick horizontal bars the 67% credible intervals, and thin bars the 95% credible intervals. Effects whose 95% credible intervals do not cross zero are interpreted as being 'significant'. Coefficient estimates are coloured by variable grouping: purple - land use; green - surrounding natural habitat; light orange - pesticide toxicity; yellow - duration of substantial landscape human habitat modification; dark orange - realized thermal niche position and effect on this of climate change. Text in parentheses refers to whether the relationship is for sites with natural local habitat or with human land use. For the baseline thermal niche position, the Baseline Thermal Position and Baseline Thermal Position^2 coefficients refer to the linear and quadratic components of the polynomial relationship, respectively, combined to describe the curivlinear relationship shown in Figure 2."}
draws <- as_draws_df(finalModel)[,3:15]

draws$`b_LandUseHuman:polyNaturalHabitat2k1` <-
  draws$`b_LandUseHuman:polyNaturalHabitat2k1` + draws$b_polyNaturalHabitat2k1
draws$`b_LandUseHuman:polyLogPestToxLow1` <- 
  draws$`b_LandUseHuman:polyLogPestToxLow1` + draws$b_polyLogPestToxLow1
draws$`b_LandUseHuman:polyAgeConv1` <- 
  draws$`b_LandUseHuman:polyAgeConv1` + draws$b_polyAgeConv1
draws$`b_LandUseHuman:polyTEI_BL21` <- 
  draws$`b_LandUseHuman:polyTEI_BL21` + draws$b_polyTEI_BL21
draws$`b_LandUseHuman:polyTEI_BL22` <- 
  draws$`b_LandUseHuman:polyTEI_BL22` + draws$b_polyTEI_BL22
draws$`b_LandUseHuman:polyTEI_delta1` <- 
  draws$`b_LandUseHuman:polyTEI_delta1` + draws$b_polyTEI_delta1

draw.means <- apply(X = draws,MARGIN = 2,FUN = mean)
draw.lower.95 <- apply(X = draws,MARGIN = 2,FUN = quantile,probs = 0.025)
draw.upper.95 <- apply(X = draws,MARGIN = 2,FUN = quantile,probs = 0.975)
draw.lower.67 <- apply(X = draws,MARGIN = 2,FUN = quantile,probs = 0.16667)
draw.upper.67 <- apply(X = draws,MARGIN = 2,FUN = quantile,probs = 0.83333)

draw.summary <- data.frame(Variable = names(draw.means),
                           Mean = draw.means,
                           Lower95 = draw.lower.95,
                           Upper95 = draw.upper.95,
                           Lower67 = draw.lower.67,
                           Upper67 = draw.upper.67)

draw.summary$Variable <- gsub("b_","",draw.summary$Variable)
draw.summary$Variable <- gsub("poly","",draw.summary$Variable)
draw.summary$Variable <- gsub("2k1","",draw.summary$Variable)
draw.summary$Variable <- gsub("LogPestToxLow1","PesticideToxicity",draw.summary$Variable)
draw.summary$Variable <- gsub("AgeConv1","ConversionDuration",draw.summary$Variable)
draw.summary$Variable <- gsub("TEI_BL21","ThermalPosition_Baseline",draw.summary$Variable)
draw.summary$Variable <- gsub("TEI_BL22","ThermalPosition_Baseline2",draw.summary$Variable)
draw.summary$Variable <- gsub("TEI_delta1","ThermalPosition_Delta",draw.summary$Variable)
# draw.summary$Variable <- gsub("LogElevation1","Elevation",draw.summary$Variable)

draw.summary$Variable <- factor(draw.summary$Variable,
                                levels = c("LandUseHuman:ThermalPosition_Delta",
                                           "ThermalPosition_Delta",
                                           "LandUseHuman:ThermalPosition_Baseline2",
                                           "LandUseHuman:ThermalPosition_Baseline",
                                           "ThermalPosition_Baseline2",
                                           "ThermalPosition_Baseline",
                                           "LandUseHuman:ConversionDuration",
                                           "ConversionDuration",
                                           "LandUseHuman:PesticideToxicity",
                                           "PesticideToxicity",
                                           "LandUseHuman:NaturalHabitat",
                                           "NaturalHabitat",
                                           "LandUseHuman"))

draw.summary$Label <- dplyr::recode(draw.summary$Variable,
                                    "LandUseHuman" = "Human Land Use",
                                    "NaturalHabitat" = "Landscape Natural Habitat (Natural)",
                                    "LandUseHuman:NaturalHabitat" = "Landscape Natural Habitat (Human)",
                                    "PesticideToxicity" = "Pesticide Toxicity (Natural)",
                                    "LandUseHuman:PesticideToxicity" = "Pesticide Toxicity (Human)",
                                    "ConversionDuration" = "Landscape Conversion Duration (Natural)",
                                    "LandUseHuman:ConversionDuration" = "Landscape Conversion Duration (Human)",
                                    "ThermalPosition_Baseline" = "Baseline Thermal Position (Natural)",
                                    "ThermalPosition_Baseline2" = "Baseline Thermal Position^2 (Natural)",
                                    "ThermalPosition_Delta" = "Thermal Position Change (Natural)",
                                    "LandUseHuman:ThermalPosition_Baseline" = "Baseline Thermal Position (Human)",
                                    "LandUseHuman:ThermalPosition_Baseline2" = "Baseline Thermal Position^2 (Human)",
                                    "LandUseHuman:ThermalPosition_Delta" = "Thermal Position Change (Human)")

draw.summary$Group <- dplyr::recode(draw.summary$Variable,
                                    "LandUseHuman" = "Land Use",
                                    "NaturalHabitat" = "Landscape Natural Habitat",
                                    "LandUseHuman:NaturalHabitat" = "Landscape Natural Habitat",
                                    "PesticideToxicity" = "Pesticide Toxicity",
                                    "LandUseHuman:PesticideToxicity" = "Pesticide Toxicity",
                                    "ConversionDuration" = "Landscape Conversion Duration",
                                    "LandUseHuman:ConversionDuration" = "Landscape Conversion Duration",
                                    "ThermalPosition_Baseline" = "Thermal Niche Properties",
                                    "ThermalPosition_Baseline2" = "Thermal Niche Properties",
                                    "ThermalPosition_Delta" = "Thermal Niche Properties",
                                    "LandUseHuman:ThermalPosition_Baseline" = "Thermal Niche Properties",
                                    "LandUseHuman:ThermalPosition_Baseline2" = "Thermal Niche Properties",
                                    "LandUseHuman:ThermalPosition_Delta" = "Thermal Niche Properties")

plotCols <- c("#D55E00","#F0E442","#E69F00","#009E73","#CC79A7")

forestPlot <- ggplot() + geom_linerangeh(mapping = aes(y=draw.summary$Label,
                                        xmin = draw.summary$Lower95, 
                                        xmax = draw.summary$Upper95,
                                        colour = draw.summary$Group)) + 
  geom_pointrangeh(mapping = aes(y = draw.summary$Label,
                                 x = draw.summary$Mean,
                                 xmin = draw.summary$Lower67,
                                 xmax = draw.summary$Upper67,
                                 colour = draw.summary$Group),
                   fatten=1.8,size=1) +
  scale_colour_manual(values = plotCols) +
  geom_vline(mapping = aes(xintercept = 0)) +
  labs(x = "Effect size",y = element_blank()) +
  guides(colour = guide_legend(nrow = 3,reverse = TRUE,
                               override.aes = aes(size=0.5))) + 
  theme_classic() +
  labs(colour = "Variable Group") + 
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10))
  
forestPlot
```

Landscape natural habitat availability had a significant positive effect on occurrence probability in human-modified habitats, but no clear effect within natural habitats (Figure \@ref(fig:figmodeleffects)a). In both natural and human-modified habitats, bumble bees are less likely to occur in landscapes with a longer duration of substantial human habitat modification (Figure \@ref(fig:figmodeleffects)b). In natural habitats, the probability of occurrence of bumble bees decreases strongly with pesticide toxicity, whereas in human-modified habitats there is a negligible effect (Figure 2c).

<!-- PLOTTING: Model conditional effects -->
```{r figmodeleffects, echo=FALSE, warning=FALSE, fig.width=17.5/2.54, fig.height=22/2.54,out.width="100%", dpi=1200, fig.cap="**Conditional modelled effects of landscape natural habitat (a), duration of landscape modification (b), pesticide toxicity (c), baseline thermal niche position (d) and change in thermal niche position from climate change (e) on the probability of occurrence of bumble bee species**. In all panels, lines represent median estimates of conditional effects, for natural (blue) or human-modified (red) habitats, more opaque shading (in the same colours) represents the 67% credible intervals, and more transparent shading the 95% credible intervals. Variables not being considered are held at their median values. Relationships are plotted for the central 95% of values sampled within each land use for each explanatory variable"}

# Set the quantiles beyond which not to plot values un-represented in the
# modelling data
exclQuantiles <- c(0.025,0.975)

#### Plotting (natural habitat) ####

# Get the quantile values of forest cover for each land use beyond which to
# exclude predictions not supported by the underlying data
QNat <- quantile(x = modelData$NaturalHabitat2k[
  modelData$LandUse=="Natural"],
  probs = exclQuantiles)
QHum <- quantile(x = modelData$NaturalHabitat2k[
  modelData$LandUse=="Human"],
  probs = exclQuantiles)

# Create data frame with variable values at which to estimate conditional effects
# Estimate conditional effects separately for median values of the variables
# in each land-use type
nd <- expand.grid(LandUse=factor(c("Natural","Human"),levels=levels(modelData$LandUse)))
nd$LogPestToxLow <- c(median(modelData$LogPestToxLow[modelData$LandUse=="Natural"]),
                      median(modelData$LogPestToxLow[modelData$LandUse=="Human"]))
nd$AgeConv <- c(median(modelData$AgeConv[modelData$LandUse=="Natural"]),
                      median(modelData$AgeConv[modelData$LandUse=="Human"]))
nd$LogElevation <- c(median(modelData$LogElevation[modelData$LandUse=="Natural"]),
                      median(modelData$LogElevation[modelData$LandUse=="Human"]))
nd$TEI_BL <- c(median(modelData$TEI_BL[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_BL[modelData$LandUse=="Human"]))
nd$TEI_delta <- c(median(modelData$TEI_delta[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_delta[modelData$LandUse=="Human"]))

# Get predicted values with both 67% (~ 1 SE) and 95% credible intervals
preds.67 <- conditional_effects(x = finalModel,effects = "NaturalHabitat2k:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.67)
preds.95 <- conditional_effects(x = finalModel,effects = "NaturalHabitat2k:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.95)

# Select only the prediction rows where other variables have median values
# corresponding with the land-use type in question
preds.67 <- rbind(preds.67$`NaturalHabitat2k:LandUse`[1:2000,]
                  [preds.67$`NaturalHabitat2k:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.67$`NaturalHabitat2k:LandUse`[2001:4000,]
                  [preds.67$`NaturalHabitat2k:LandUse`[2001:4000,]$LandUse=="Human",])
preds.95 <- rbind(preds.95$`NaturalHabitat2k:LandUse`[1:2000,]
                  [preds.95$`NaturalHabitat2k:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.95$`NaturalHabitat2k:LandUse`[2001:4000,]
                  [preds.95$`NaturalHabitat2k:LandUse`[2001:4000,]$LandUse=="Human",])

# Exclude extreme predictions not supported by the underlying data
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$NaturalHabitat2k < QNat[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$NaturalHabitat2k > QNat[2]))] <- NA

preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$NaturalHabitat2k < QHum[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$NaturalHabitat2k > QHum[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$NaturalHabitat2k < QNat[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$NaturalHabitat2k > QNat[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$NaturalHabitat2k < QHum[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$NaturalHabitat2k > QHum[2]))] <- NA

# Remove NA values from the predicted values
preds.67 <- preds.67[!is.na(preds.67$estimate__),]
preds.95 <- preds.95[!is.na(preds.95$estimate__),]

plotHab <- ggplot() + 
  geom_ribbon(data = preds.95, 
              mapping = aes(x = NaturalHabitat2k/10,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.05) + 
  geom_ribbon(data = preds.67,
              mapping = aes(x = NaturalHabitat2k/10,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.15) + 
  geom_line(data = preds.67,
            mapping = aes(x = NaturalHabitat2k/10,y = estimate__,col = LandUse),
            linewidth = 1) + 
  scale_fill_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_color_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_x_continuous(name = "Natural habitat (%)",limits = c(0,100)) + 
  scale_y_continuous(name = "Probability of occurrence",limits = c(0,1)) + 
  theme_classic() + 
  theme(legend.justification = c(0.5, 0.5),
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10)) + 
  labs(colour = "Land Use", fill = "Land Use")

plotHab2 <- plotHab + theme(legend.position = "none")

#### Plotting (landscape conversion age) ####

# Get the quantile values of forest cover for each land use beyond which to
# exclude predictions not supported by the underlying data
QNat <- quantile(x = modelData$AgeConv[
  modelData$LandUse=="Natural"],
  probs = exclQuantiles)
QHum <- quantile(x = modelData$AgeConv[
  modelData$LandUse=="Human"],
  probs = exclQuantiles)

# Create data frame with variable values at which to estimate conditional effects
# Estimate conditional effects separately for median values of the variables
# in each land-use type
nd <- expand.grid(LandUse=factor(c("Natural","Human"),levels=levels(modelData$LandUse)))
nd$LogPestToxLow <- c(median(modelData$LogPestToxLow[modelData$LandUse=="Natural"]),
                      median(modelData$LogPestToxLow[modelData$LandUse=="Human"]))
nd$NaturalHabitat2k <- c(median(modelData$NaturalHabitat2k[modelData$LandUse=="Natural"]),
                      median(modelData$NaturalHabitat2k[modelData$LandUse=="Human"]))
nd$LogElevation <- c(median(modelData$LogElevation[modelData$LandUse=="Natural"]),
                      median(modelData$LogElevation[modelData$LandUse=="Human"]))
nd$TEI_BL <- c(median(modelData$TEI_BL[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_BL[modelData$LandUse=="Human"]))
nd$TEI_delta <- c(median(modelData$TEI_delta[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_delta[modelData$LandUse=="Human"]))

# Get predicted values with both 67% (~ 1 SE) and 95% credible intervals
preds.67 <- conditional_effects(x = finalModel,effects = "AgeConv:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.67)
preds.95 <- conditional_effects(x = finalModel,effects = "AgeConv:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.95)

# Select only the prediction rows where other variables have median values
# corresponding with the land-use type in question
preds.67 <- rbind(preds.67$`AgeConv:LandUse`[1:2000,]
                  [preds.67$`AgeConv:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.67$`AgeConv:LandUse`[2001:4000,]
                  [preds.67$`AgeConv:LandUse`[2001:4000,]$LandUse=="Human",])
preds.95 <- rbind(preds.95$`AgeConv:LandUse`[1:2000,]
                  [preds.95$`AgeConv:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.95$`AgeConv:LandUse`[2001:4000,]
                  [preds.95$`AgeConv:LandUse`[2001:4000,]$LandUse=="Human",])

# Exclude extreme predictions not supported by the underlying data
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$AgeConv < QNat[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$AgeConv > QNat[2]))] <- NA

preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$AgeConv < QHum[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$AgeConv > QHum[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$AgeConv < QNat[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$AgeConv > QNat[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$AgeConv < QHum[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$AgeConv > QHum[2]))] <- NA

# Remove NA values from the predicted values
preds.67 <- preds.67[!is.na(preds.67$estimate__),]
preds.95 <- preds.95[!is.na(preds.95$estimate__),]

plotAge <- ggplot() + 
  geom_ribbon(data = preds.95, 
              mapping = aes(x = AgeConv,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.05) + 
  geom_ribbon(data = preds.67,
              mapping = aes(x = AgeConv,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.15) + 
  geom_line(data = preds.67,
            mapping = aes(x = AgeConv,y = estimate__,col = LandUse),
            linewidth = 1) + 
  scale_fill_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_color_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_x_continuous(name = "Duration of landscape conversion (yrs)") + 
  scale_y_continuous(name = "Probability of occurrence",limits = c(0,1)) + 
  theme_classic() + 
  theme(legend.position = "none",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10))


#### Plotting (pesticide toxicity) ####

# Get the quantile values of forest cover for each land use beyond which to
# exclude predictions not supported by the underlying data
QNat <- quantile(x = modelData$LogPestToxLow[
  modelData$LandUse=="Natural"],
  probs = exclQuantiles)
QHum <- quantile(x = modelData$LogPestToxLow[
  modelData$LandUse=="Human"],
  probs = exclQuantiles)

# Create data frame with variable values at which to estimate conditional effects
# Estimate conditional effects separately for median values of the variables
# in each land-use type
nd <- expand.grid(LandUse=factor(c("Natural","Human"),levels=levels(modelData$LandUse)))
nd$AgeConv <- c(median(modelData$AgeConv[modelData$LandUse=="Natural"]),
                      median(modelData$AgeConv[modelData$LandUse=="Human"]))
nd$NaturalHabitat2k <- c(median(modelData$NaturalHabitat2k[modelData$LandUse=="Natural"]),
                      median(modelData$NaturalHabitat2k[modelData$LandUse=="Human"]))
nd$LogElevation <- c(median(modelData$LogElevation[modelData$LandUse=="Natural"]),
                      median(modelData$LogElevation[modelData$LandUse=="Human"]))
nd$TEI_BL <- c(median(modelData$TEI_BL[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_BL[modelData$LandUse=="Human"]))
nd$TEI_delta <- c(median(modelData$TEI_delta[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_delta[modelData$LandUse=="Human"]))

# Get predicted values with both 67% (~ 1 SE) and 95% credible intervals
preds.67 <- conditional_effects(x = finalModel,effects = "LogPestToxLow:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.67)
preds.95 <- conditional_effects(x = finalModel,effects = "LogPestToxLow:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.95)

# Select only the prediction rows where other variables have median values
# corresponding with the land-use type in question
preds.67 <- rbind(preds.67$`LogPestToxLow:LandUse`[1:2000,]
                  [preds.67$`LogPestToxLow:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.67$`LogPestToxLow:LandUse`[2001:4000,]
                  [preds.67$`LogPestToxLow:LandUse`[2001:4000,]$LandUse=="Human",])
preds.95 <- rbind(preds.95$`LogPestToxLow:LandUse`[1:2000,]
                  [preds.95$`LogPestToxLow:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.95$`LogPestToxLow:LandUse`[2001:4000,]
                  [preds.95$`LogPestToxLow:LandUse`[2001:4000,]$LandUse=="Human",])

# Exclude extreme predictions not supported by the underlying data
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$LogPestToxLow < QNat[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$LogPestToxLow > QNat[2]))] <- NA

preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$LogPestToxLow < QHum[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$LogPestToxLow > QHum[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$LogPestToxLow < QNat[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$LogPestToxLow > QNat[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$LogPestToxLow < QHum[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$LogPestToxLow > QHum[2]))] <- NA

# Remove NA values from the predicted values
preds.67 <- preds.67[!is.na(preds.67$estimate__),]
preds.95 <- preds.95[!is.na(preds.95$estimate__),]

plotPest <- ggplot() + 
  geom_ribbon(data = preds.95, 
              mapping = aes(x = exp(LogPestToxLow)-1,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.05) + 
  geom_ribbon(data = preds.67,
              mapping = aes(x = exp(LogPestToxLow)-1,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.15) + 
  geom_line(data = preds.67,
            mapping = aes(x = exp(LogPestToxLow)-1,y = estimate__,col = LandUse),
            linewidth = 1) + 
  scale_fill_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_color_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_x_continuous(name = "Pesticide toxicity") + 
  scale_y_continuous(name = "Probability of occurrence",limits = c(0,1)) + 
  theme_classic() + 
  theme(legend.position = "none",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10))

#### Plotting (Baseline TEI) ####

# Get the quantile values of forest cover for each land use beyond which to
# exclude predictions not supported by the underlying data
QNat <- quantile(x = modelData$TEI_BL[
  modelData$LandUse=="Natural"],
  probs = exclQuantiles)
QHum <- quantile(x = modelData$TEI_BL[
  modelData$LandUse=="Human"],
  probs = exclQuantiles)

# Create data frame with variable values at which to estimate conditional effects
# Estimate conditional effects separately for median values of the variables
# in each land-use type
nd <- expand.grid(LandUse=factor(c("Natural","Human"),levels=levels(modelData$LandUse)))
nd$AgeConv <- c(median(modelData$AgeConv[modelData$LandUse=="Natural"]),
                      median(modelData$AgeConv[modelData$LandUse=="Human"]))
nd$NaturalHabitat2k <- c(median(modelData$NaturalHabitat2k[modelData$LandUse=="Natural"]),
                      median(modelData$NaturalHabitat2k[modelData$LandUse=="Human"]))
nd$LogElevation <- c(median(modelData$LogElevation[modelData$LandUse=="Natural"]),
                      median(modelData$LogElevation[modelData$LandUse=="Human"]))
nd$LogPestToxLow <- c(median(modelData$LogPestToxLow[modelData$LandUse=="Natural"]),
                      median(modelData$LogPestToxLow[modelData$LandUse=="Human"]))
nd$TEI_delta <- c(median(modelData$TEI_delta[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_delta[modelData$LandUse=="Human"]))

# Get predicted values with both 67% (~ 1 SE) and 95% credible intervals
preds.67 <- conditional_effects(x = finalModel,effects = "TEI_BL:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.67)
preds.95 <- conditional_effects(x = finalModel,effects = "TEI_BL:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.95)

# Select only the prediction rows where other variables have median values
# corresponding with the land-use type in question
preds.67 <- rbind(preds.67$`TEI_BL:LandUse`[1:2000,]
                  [preds.67$`TEI_BL:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.67$`TEI_BL:LandUse`[2001:4000,]
                  [preds.67$`TEI_BL:LandUse`[2001:4000,]$LandUse=="Human",])
preds.95 <- rbind(preds.95$`TEI_BL:LandUse`[1:2000,]
                  [preds.95$`TEI_BL:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.95$`TEI_BL:LandUse`[2001:4000,]
                  [preds.95$`TEI_BL:LandUse`[2001:4000,]$LandUse=="Human",])

# Exclude extreme predictions not supported by the underlying data
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$TEI_BL < QNat[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$TEI_BL > QNat[2]))] <- NA

preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$TEI_BL < QHum[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$TEI_BL > QHum[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$TEI_BL < QNat[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$TEI_BL > QNat[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$TEI_BL < QHum[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$TEI_BL > QHum[2]))] <- NA

# Remove NA values from the predicted values
preds.67 <- preds.67[!is.na(preds.67$estimate__),]
preds.95 <- preds.95[!is.na(preds.95$estimate__),]

plotTEI_bl <- ggplot() + 
  geom_ribbon(data = preds.95, 
              mapping = aes(x = TEI_BL,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.05) + 
  geom_ribbon(data = preds.67,
              mapping = aes(x = TEI_BL,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.15) + 
  geom_line(data = preds.67,
            mapping = aes(x = TEI_BL,y = estimate__,col = LandUse),
            linewidth = 1) + 
  scale_fill_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_color_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_x_continuous(name = "Baseline thermal position") + 
  scale_y_continuous(name = "Probability of occurrence",limits = c(0,1)) + 
  theme_classic() + 
  theme(legend.position = "none",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10))

#### Plotting (delta TEI) ####

# Get the quantile values of forest cover for each land use beyond which to
# exclude predictions not supported by the underlying data
QNat <- quantile(x = modelData$TEI_delta[
  modelData$LandUse=="Natural"],
  probs = exclQuantiles)
QHum <- quantile(x = modelData$TEI_delta[
  modelData$LandUse=="Human"],
  probs = exclQuantiles)

# Create data frame with variable values at which to estimate conditional effects
# Estimate conditional effects separately for median values of the variables
# in each land-use type
nd <- expand.grid(LandUse=factor(c("Natural","Human"),levels=levels(modelData$LandUse)))
nd$AgeConv <- c(median(modelData$AgeConv[modelData$LandUse=="Natural"]),
                      median(modelData$AgeConv[modelData$LandUse=="Human"]))
nd$NaturalHabitat2k <- c(median(modelData$NaturalHabitat2k[modelData$LandUse=="Natural"]),
                      median(modelData$NaturalHabitat2k[modelData$LandUse=="Human"]))
nd$LogElevation <- c(median(modelData$LogElevation[modelData$LandUse=="Natural"]),
                      median(modelData$LogElevation[modelData$LandUse=="Human"]))
nd$LogPestToxLow <- c(median(modelData$LogPestToxLow[modelData$LandUse=="Natural"]),
                      median(modelData$LogPestToxLow[modelData$LandUse=="Human"]))
nd$TEI_BL <- c(median(modelData$TEI_BL[modelData$LandUse=="Natural"]),
                      median(modelData$TEI_BL[modelData$LandUse=="Human"]))

# Get predicted values with both 67% (~ 1 SE) and 95% credible intervals
preds.67 <- conditional_effects(x = finalModel,effects = "TEI_delta:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.67)
preds.95 <- conditional_effects(x = finalModel,effects = "TEI_delta:LandUse",
                                resolution = 1000,
                                conditions = nd,plot = FALSE,prob=0.95)

# Select only the prediction rows where other variables have median values
# corresponding with the land-use type in question
preds.67 <- rbind(preds.67$`TEI_delta:LandUse`[1:2000,]
                  [preds.67$`TEI_delta:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.67$`TEI_delta:LandUse`[2001:4000,]
                  [preds.67$`TEI_delta:LandUse`[2001:4000,]$LandUse=="Human",])
preds.95 <- rbind(preds.95$`TEI_delta:LandUse`[1:2000,]
                  [preds.95$`TEI_delta:LandUse`[1:2000,]$LandUse=="Natural",],
                  preds.95$`TEI_delta:LandUse`[2001:4000,]
                  [preds.95$`TEI_delta:LandUse`[2001:4000,]$LandUse=="Human",])

# Exclude extreme predictions not supported by the underlying data
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$TEI_delta < QNat[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Natural") & 
                            (preds.67$TEI_delta > QNat[2]))] <- NA

preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$TEI_delta < QHum[1]))] <- NA
preds.67$estimate__[which((preds.67$LandUse=="Human") & 
                            (preds.67$TEI_delta > QHum[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$TEI_delta < QNat[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Natural") & 
                            (preds.95$TEI_delta > QNat[2]))] <- NA

preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$TEI_delta < QHum[1]))] <- NA
preds.95$estimate__[which((preds.95$LandUse=="Human") & 
                            (preds.95$TEI_delta > QHum[2]))] <- NA

# Remove NA values from the predicted values
preds.67 <- preds.67[!is.na(preds.67$estimate__),]
preds.95 <- preds.95[!is.na(preds.95$estimate__),]

plotTEI_delta <- ggplot() + 
  geom_ribbon(data = preds.95, 
              mapping = aes(x = TEI_delta,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.05) + 
  geom_ribbon(data = preds.67,
              mapping = aes(x = TEI_delta,ymin = lower__,
                            ymax = upper__,fill = LandUse),alpha = 0.15) + 
  geom_line(data = preds.67,
            mapping = aes(x = TEI_delta,y = estimate__,col = LandUse),
            linewidth = 1) + 
  scale_fill_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_color_manual(values = c("#2F67B1","#BF2C23")) + 
  scale_x_continuous(name = "\u0394 thermal position") + 
  scale_y_continuous(name = "Probability of occurrence",limits = c(0,1)) + 
  theme_classic() + 
  theme(legend.position = "none",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 10))

legend <- get_legend(plotHab)

plots <- plot_grid(plotHab2,plotAge,plotPest,plotTEI_bl,plotTEI_delta,legend,
                   nrow = 3,ncol = 2,
                   labels =c("a)","b)","c)","d)","e)"),
                   label_x = -0.02,label_size = 12,label_fontface = "plain")

plots

```

Populations near the centre of species' thermal  niches (the exact centre of the niche has a thermal position of 0.5) showed the greatest reduction in probability of occurrence in human-modified compared to natural habitats, whereas for populations near the upper thermal niche limit, occurrence probability was similar in natural and human-modified habitats (Figure \@ref(fig:figmodeleffects)d). Increases in realized thermal niche position caused by climate warming were associated with a weak (non-significant) negative effect on occurrence probability in natural habitats, but a positive effect in human-modified habitats (Figure \@ref(fig:figforest); Figure \@ref(fig:figmodeleffects)e).

```{r autocorrel_test, echo=FALSE, warning=FALSE}
siteAvgResids <- data.frame(resids=tapply(X = residuals(finalModel)[,1],
                                          INDEX = droplevels(modelData)$SSBS,
                                          FUN = mean))
siteAvgResids$SSBS <- row.names(siteAvgResids)
row.names(siteAvgResids) <- NULL
siteAvgResids$Longitude <- modelData$Longitude[
  match(siteAvgResids$SSBS,modelData$SSBS)]
siteAvgResids$Latitude <- modelData$Latitude[
  match(siteAvgResids$SSBS,modelData$SSBS)]
siteAvgResids$SS <- modelData$SS[
  match(siteAvgResids$SSBS,modelData$SSBS)]

ds.nb <- dnearneigh(cbind(siteAvgResids$Longitude,siteAvgResids$Latitude),
                    d1=0.00000001,d2=10)

ds.listw <- nb2listw(ds.nb)

mt <- moran.test(siteAvgResids$resids,ds.listw)

moran.i <- round(mt$estimate[1],4)
moran.p <- round(mt$p.value,2)

```

Individual model chains converged well (all Rhat &le; `r max.rhat`; Figure S4). There was a good correspondence between model predicted values and observations (Figure S5). A Moran's test across all model residuals did not reveal significant spatial autocorrelation (I = `r moran.i`; P = `r moran.p`). However, there was significant spatial autocorrelation (P < 0.05) in the residuals associated with 13.6% of the underlying studies in the dataset, higher than the 5% expected by chance (Figure S6). Importantly, the records from studies with significant residual spatial autocorrelation did not occupy extreme values of any explanatory variable (Figure S3), and so bias in coefficient estimates is unlikely. Results were very similar when landscape natural-habitat availability was estimated at 1-km or 5-km spatial grain (Figures S7 & S8), and when using high instead of low estimates of cumulative pesticide toxicity (Figure S11). Estimates of the effect of duration of substantial habitat modification were weaker (although qualitatively unchanged) when substantial habitat modification was defined using a threshold of 10% or 50% (rather than 30%) conversion of landscapes to human-modified land uses (Figures S9 & S10). Results were also robust when fitting a zero-inflated negative binomial model of relative abundance, the key differences being that effects of natural habitat were somewhat weaker, and effects of pesticide toxicity stronger (Figure S12).

```{r preds_sites, echo=FALSE,results=FALSE}

#### Predictions for sampled locations ####

# Create baseline data-frame with all natural land use, 100% surrounding natural
# habitat (note that 100% is stored as 1,000), no pesticide and no climate change
nd.bl <- droplevels(modelData)
nd.bl$LandUse <- factor("Natural",levels=c("Natural","Human"))
nd.bl$NaturalHabitat2k <- 1000
nd.bl$LogPestToxLow <- 0
nd.bl$AgeConv <- 0
nd.bl$TEI_delta <- 0

# Extract observed data-frame
nd.obs <- droplevels(modelData)

# cl <- makeCluster(detectCores()-1)
cl <- makeCluster(8)
clusterExport(cl = cl,varlist = c("finalModel","nd.bl","nd.obs"))

# Make 1,000 predictions of % difference between observed
# and baseline conditions
preds <- parSapply(cl = cl,X = 1:1000,FUN = function(i){
  
  # Draw one of the posterior parameter estimates at random
  i <- sample(x = 1:4000,size = 1,replace = FALSE)
  
  # Use this set of posterior parameter estimates to get predicted
  # probability of occurrence under baseline and observed environmental
  # conditions
  y.bl <- brms::posterior_epred(
    object = finalModel,newdata = nd.bl,re_formula = NA,
    draw_ids = i)[1,]
  
  y.obs <- brms::posterior_epred(
    object = finalModel,newdata = nd.obs,re_formula = NA,
    draw_ids = i)[1,]
  
  y <- ((y.obs/y.bl)*100)-100
  
  return(y)
})

stopCluster(cl)

# Calculate median and credible intervals across replicate predictions
y.median <- apply(X = preds,MARGIN = 1,FUN = median)
y.lower.95 <- apply(X = preds,MARGIN = 1,FUN = quantile,probs=0.025)
y.upper.95 <- apply(X = preds,MARGIN = 1,FUN = quantile,probs=0.975)
y.lower.67 <- apply(X = preds,MARGIN = 1,FUN = quantile,probs=0.16667)
y.upper.67 <- apply(X = preds,MARGIN = 1,FUN = quantile,probs=0.83333)

# Summarize median, and credible intervals of predictions
# across the two land-use types
lu <- factor(c("Natural","Human"),levels=c("Natural","Human"))
lu.means.median <- tapply(X = y.median,INDEX = nd.obs$LandUse,FUN = median)
lu.means.lower.95 <- tapply(X = y.lower.95,INDEX = nd.obs$LandUse,FUN = median)
lu.means.upper.95 <- tapply(X = y.upper.95,INDEX = nd.obs$LandUse,FUN = median)
lu.means.lower.67 <- tapply(X = y.lower.67,INDEX = nd.obs$LandUse,FUN = median)
lu.means.upper.67 <- tapply(X = y.upper.67,INDEX = nd.obs$LandUse,FUN = median)

```

Compared to estimated occurrence probability under hypothetical reference conditions (100% surrounding natural habitat, and no history of human land use, application of agricultural chemicals or climate change), we estimate that bumble bee probability of occurrence is lower by `r -round(lu.means.median['Natural'],0)`% (95% CI: `r -round(lu.means.upper.95['Natural'],0)`% &ndash; `r -round(lu.means.lower.95['Natural'],0)`%) across sampled natural habitats, and by `r -round(lu.means.median['Human'],0)`% (95% CI: `r -round(lu.means.upper.95['Human'],0)`% &ndash; `r -round(lu.means.lower.95['Human'],0)`%) across sampled human land uses (Figure \@ref(fig:figpredssites)).

```{r figpredssites, echo=FALSE, warning=FALSE, fig.width=17.5/2.54, fig.height=18/2.54,out.width="100%", dpi=1200, fig.cap="**Model-inferred differences in average bumble bee occurrence probability at sampled sites in natural and human-modified habitats, compared to hypothetical reference conditions (natural land use across the whole landscape, no human land-use history, no agricultural chemical application, and no climate change)**. Predictions were generated for every species recorded at each sampled site across both natural and human-modified habitats, based on the estimates of the explanatory variables. Points show the median prediction, while thick and thin error bars represent the 67% and 95% confidence intervals, respectively, based on sampling 1,000 coefficient estimates from the model posteriors."}

# Average predicted differences in occurrence probability across sampled sites
y.median.sites <- tapply(X = y.median,INDEX = nd.obs$SSBS,FUN = mean)
y.upper.sites <- tapply(X = y.upper.95,INDEX = nd.obs$SSBS,FUN = mean)
y.lower.sites <- tapply(X = y.lower.95,INDEX = nd.obs$SSBS,FUN = mean)

# Create data frame for site-level averages
sites <- data.frame(y=y.median.sites,yplus=y.upper.sites,yminus=y.lower.sites,
                    SSBS=names(y.median.sites))
sites$Longitude <- nd.obs$Longitude[match(sites$SSBS,nd.obs$SSBS)]
sites$Latitude <- nd.obs$Latitude[match(sites$SSBS,nd.obs$SSBS)]

# Create basemap for North American and Western European sub-regions
baseMap <- vect("0_data/UN_subregion.shp")
baseMap <- terra::subset(baseMap,baseMap$SUBREGION %in% c(21,13,154,155,39))
baseMap <- st_as_sf(baseMap)

# Create map of predictions for sites sampled in PREDICTS
predsMap <- vect(x = sites,geom=c("Longitude","Latitude"),
                 crs = '+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0')
predsMap <- st_as_sf(predsMap)
# Add column showing land use
predsMap$LandUse <- modelData$LandUse[match(predsMap$SSBS,modelData$SSBS)]

# Plot base map and predictions for sampled locations
mapPreds <- ggplot() + geom_sf(data = baseMap,fill = "#ffffff") + 
  geom_sf(data=predsMap,mapping = aes(colour=y,shape=LandUse)) + 
  scale_color_continuous("\u0394 Probability of\noccurrence (%)",type = "viridis") + 
  scale_x_continuous(limits = c(-180,60)) + 
  theme_classic() + 
  theme(axis.text = element_text(size = 10),
        legend.title = element_text(size = 12),
        legend.text = element_text(size=10))

# Create error-bar plot showing predictions at sampled locations
# by land-use type
plotLU <- ggplot() + 
  geom_linerange(mapping = aes(x = lu,
                               ymin = lu.means.lower.95,
                               ymax = lu.means.upper.95,
                               col = lu)) +
  geom_pointrange(mapping = aes(x = lu,
                                y = lu.means.median,
                                ymin = lu.means.lower.67,
                                ymax = lu.means.upper.67,
                                col = lu),
                  linewidth = 1, size = 0.8) +
  scale_colour_manual(values = c("#2F67B1","#BF2C23")) +
  scale_y_continuous(name = "\u0394 Probability of occurrence (%)",
                     limits = c(-100,0)) + 
  scale_x_discrete(name = "Land use") + 
  geom_hline(mapping = aes(yintercept = 0),
             alpha = 0.3, linetype = 'dashed') +
  theme_classic() + 
  theme(legend.position = "none",
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 12))

# Create final figure
plots <- cowplot::plot_grid(mapPreds,plotLU,nrow = 2,ncol = 1,
                   rel_heights = c(0.5,0.5),rel_widths = c(1,0.1))
plots
```

# Discussion

Our results reveal important interactive effects of climate change, landscape habitat modification and local land-use conditions on bumble bee occurrence probability. Climate-land-use interactions may exacerbate biodiversity change [@Oliver2014; @Williams2020], but relatively few studies have focused on insects [@Hamblin2017; @Oliver2017; @Fourcade2019; @Waldock2020; @Kammerer2021; @Outhwaite2022; @Janousek2023; @Hemberger2024]. We estimate that bumble bee occurrence probability has been reduced substantially across both natural and human-modified habitats as a result of human pressures.

We find that bumble bee occurrence probability is reduced in human-modified habitats in landscapes with lower cover of natural habitats, and across all habitats in landscapes that have been substantially modified by humans for a longer time. Our findings are consistent with previous, local or regional-scale studies, which generally found negative effects of human land use, and positive effects of landscape semi-natural habitat on bumble bees [@Samuelson2018; @Proesmans2019; @Janousek2023]. We show the generality of these effects across two north-temperate continents, and show for the first time the importance of landscape land-use history in shaping responses.

Interestingly, occurrence probability is reduced most by human land use in the centre of species' thermal niches (indicated by a greater difference in occurrence probability between natural and human-modified habitats; Figure \@ref(fig:figmodeleffects)d), in contrast with previous research both on vertebrates [@Williams2021] and on insects [@Waldock2020]. This may arise because the generally higher occurrence frequency near the niche centre means that factors other than climate play an important role in shaping species presence or absence. For example, floral resources are essential for bumble bees [@Requier2020; @McCombs2022], and are known to be limiting in some high-intensity agricultural landscapes, particularly at certain times of year [@Timberlake2019].

Climate warming, which moves populations closer to species' upper thermal limits, slightly reduced occurrence probability in natural habitats, but increased occurrence probability in human land uses. This contrasts with the results of a study on birds in North America, where declines in response to habitat loss were strongest where summer temperatures had warmed most [@Northrup2019]. This finding may indicate that bumble bee species found within human-modified habitats constitute a filtered set that are relatively more resilient to intensifying pressures such as climate change. Overall, the interactive effects of climate change and land use appear more complicated for bumble bees than for other groups of species, and it is important to note that varying rates of warming in different locations will add further complexity.

The cumulative toxicity of pesticides applied in landscapes is associated with steep reductions in occurrence probability of bumble bees in natural habitats (often these natural habitats are in close proximity to farmland), but surprisingly not within agricultural areas. The absence of an effect of pesticide toxicity in agricultural areas could again indicate that bumble bee assemblages in these areas are already filtered to include only those species relatively more tolerant of human pressures. Indeed, a study of UK bees found detectable, but weak, effects of pesticide (specifically neonicotinoid) exposure on population trends within farmed landscapes [@Woodcock2016]. Nevertheless, we caution that the available global data on application of agricultural chemicals are resolved only at a coarse spatial resolution of approximately 10 km [@Mueller2012; @Maggi2019], and future studies will benefit from more finely resolved estimates of pesticide application [@Mesnage2021].

Using our models to infer changes in bumble bee occurrence probability as a result of cumulative pressures suggests that landscape modification, combined with recent climate change, has reduced bumble bee occurrence probability within natural habitats almost as much as in human-modified areas. While agricultural areas experience the direct impacts of habitat loss, many natural habitats are exposed to degradation of their surrounding landscapes. Additionally, both natural and human-modified habitats are exposed to the effects of climate change. Land-use impacts are often quantified using spatial analyses, comparing sampled biodiversity between natural and modified areas [@Newbold2015]. The fact that natural habitats are composed of depauperate communities, heavily impacted by changes in their landscapes, means that such estimates likely underestimate land-use impacts substantially [@Newbold2019].

Understanding changes in biodiversity through time using a spatial analysis is inevitably subject to limitations. Most importantly, spatial analyses cannot consider potential time-lagged effects of environmental changes [@DePalma2018]. However, time-series data for insects are lacking in most regions. Second, there is likely important interspecific variation in responses [@Hemberger2021], but the database we used did not have sufficient sampling of individual species to consider this. Third, large-scale studies typically rely on coarsely resolved estimates of pressures. While the availability of fine-scale land-use/land-cover data has improved rapidly in recent years [@Jung2020], we still lack fine-scale estimates of pesticide application, a particularly important pressure for flying insects [@Wagner2021]. The classification of local land-use type we used is also thematically coarse, simply dividing sites into natural or human-modified habitats. Our models revealed relatively little difference overall in probability of occurrence between these land-use types, although there were important interactive effects with the other explanatory variables. To estimate the position of populations within species' thermal niches, we use information on species' realized distributions with respect to temperature. Such realized temperature limits may not correspond with the fundamental physiological limits of species, but estimates of the latter [@Bennett2018] are available for too few species to be useful in large-scale analyses. A previous study showed that realized thermal limits are clearly associated with occupancy change in response to recent climate change [@Soroye2020], suggesting these limits are ecologically important. We detected significant spatial autocorrelation in the residuals associated with a slightly higher fraction of studies than expected by chance, which may lead to some uncertainty in the modelled effects, although the fact that studies whose residuals showed significant spatial autocorrelation did not occupy extreme values of any of the explanatory variables suggests that substantial bias is unlikely. In general, correlative analyses can be consistent with, but not prove, causal mechanisms. The complexity of the interactive effects we document highlights a need for experimental studies or detailed field-level assessments to investigate the mechanistic basis for the patterns we report. Finally, it is important to highlight the entirely hypothetical reference against which we estimated potential changes in bumblebee assemblages caused by human activities. While this reference isn't expected to reflect actual environmental conditions at any time in recent history, it serves to illustrate the potential magnitude of human impacts on bumble bee biodiversity, across both natural and human-modified habitats.

In conclusion, interactions between climate change, climatic niche position and land-use pressures (contemporary and historical) drive substantial changes in bumble bee occurrence across both natural and human-modified habitats. Given the importance of bumble bees for pollinating wild plants [@Ollerton2017] and agricultural crops [@Pritchard2020], these changes are likely to have important effects on both natural ecosystems and on our ability to grow food. Ongoing changes in climate and land use are likely to lead to a further reshaping of bumble bee communities across natural and human-modified habitats. To predict future changes in biodiversity robustly, including of bumble bees, we must account for the complex interactive effects of climate change and land use.

<!--
```{r session_info, echo=FALSE,reulst=TRUE,eval=TRUE}
print(sessionInfo())
```
-->

# Data and Code Availability:

The code needed to run all of the analyses is publicly available on ([GitHub](https://github.com/timnewbold/BumblebeesLUClimStudy)). The data needed to run the code are published on FigShare, DOI: ([10.6084/m9.figshare.28713143](http://doi.org/10.6084/m9.figshare.28713143)), except for datasets for which we do not have permission to republish, but which are freely available in other public repositories. Instructions for obtaining these additional datasets are given both in the FigShare repository description, and at the top of the first two code scripts.

# Acknowledgments

This research was supported by a Royal Society International Exchanges grant to TN and JK (IE161031), UK Natural Environment Research Council grants to TN (NE/R010811/1 and NE/V006533/1) and a Royal Society University Research Fellowship to TN (UF150526). The authors declare no competing interests.

# References
